Extract SetTimeKeyMixin from mixin.rb
